Adds shift key requirement for selection
Requires the shift key to be pressed for selection and deselection actions.

This prevents unintended actions when the user is not actively trying to select or deselect text. The shift key acts as a modifier to ensure the user's intent. Resets popup if shift key is released.

Adds flag changed event to global event listener.